refactor(core): 降低核心模块复杂度并归档提案
本次提交完成 reduce-core-complexity-preserve-behavior 的行为保持型重构收口。

主要改动包括：抽取 Tauri 文本文件 bridge 到独立模块并保留旧入口导出；抽取 threads reducer 的 tool status 收敛逻辑并拆分 reasoning 回归测试；抽取 Rust runtime identity helper；抽取 Settings 实验开关展示组件；拆分 Spec Hub controls 样式并保持导入顺序；同步 core-complexity-governance 主 spec 并归档 OpenSpec change。

验证结果：openspec validate --all --strict 通过，归档前 openspec validate reduce-core-complexity-preserve-behavior --strict 通过，git diff --check / staged check 通过；人工桌面回归覆盖 App 启动、workspace 切换、Codex send/interruption/continue、thread history、settings 持久化、AGENTS/CLAUDE 读写、file preview、Git status/diff/history、Spec Hub 布局和主题，结果未发现问题。
