+++ openspec/changes/archive/2026-05-02-reduce-core-complexity-preserve-behavior/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-02
