+++ src/services/tauri.ts
@@ -57,6 +57,25 @@ export {
 } from "./tauri/sessionManagement";
 export type { CodexRuntimeReloadResult } from "./tauri/settings";
 export { getCodexConfigPath, getCodexUnifiedExecExternalStatus, reloadCodexRuntimeConfig, restoreCodexUnifiedExecOfficialDefault, setCodexUnifiedExecOfficialOverride } from "./tauri/settings";
+export type {
+  AgentMdResponse,
+  ClaudeMdResponse,
+  GlobalAgentsResponse,
+  GlobalCodexAuthResponse,
+  GlobalCodexConfigResponse,
+  TextFileResponse,
+} from "./tauri/textFiles";
+export {
+  readAgentMd,
+  readClaudeMd,
+  readGlobalAgentsMd,
+  readGlobalCodexAuthJson,
+  readGlobalCodexConfigToml,
+  writeAgentMd,
+  writeClaudeMd,
+  writeGlobalAgentsMd,
+  writeGlobalCodexConfigToml,
+} from "./tauri/textFiles";
 export { getComputerUseBridgeStatus, runComputerUseActivationProbe, runComputerUseCodexBroker, runComputerUseHostContractDiagnostics } from "./tauri/computerUse";
 export { runClaudeDoctor, runCodexDoctor } from "./tauri/doctor";
 export type {
@@ -304,48 +323,6 @@ export interface ClaudeSessionSummaryPayload {
   firstMessage: string;
   updatedAt: number;
   fileSizeBytes?: number;
-}
-
-export type TextFileResponse = {
-  exists: boolean;
-  content: string;
-  truncated: boolean;
-};
-
-export type GlobalAgentsResponse = TextFileResponse;
-export type GlobalCodexConfigResponse = TextFileResponse;
-export type GlobalCodexAuthResponse = TextFileResponse;
-export type AgentMdResponse = TextFileResponse;
-
-type FileScope = "workspace" | "global";
-type FileKind = "agents" | "claude" | "config" | "auth";
-
-async function fileRead(scope: FileScope, kind: FileKind, workspaceId?: string): Promise<TextFileResponse> {
-  return invoke<TextFileResponse>("file_read", { scope, kind, workspaceId });
-}
-
-async function fileWrite(scope: FileScope, kind: FileKind, content: string, workspaceId?: string): Promise<void> {
-  return invoke("file_write", { scope, kind, workspaceId, content });
-}
-
-export async function readGlobalAgentsMd(): Promise<GlobalAgentsResponse> {
-  return fileRead("global", "agents");
-}
-
-export async function writeGlobalAgentsMd(content: string): Promise<void> {
-  return fileWrite("global", "agents", content);
-}
-
-export async function readGlobalCodexConfigToml(): Promise<GlobalCodexConfigResponse> {
-  return fileRead("global", "config");
-}
-
-export async function writeGlobalCodexConfigToml(content: string): Promise<void> {
-  return fileWrite("global", "config", content);
-}
-
-export async function readGlobalCodexAuthJson(): Promise<GlobalCodexAuthResponse> {
-  return fileRead("global", "auth");
 }
 
 export async function getConfigModel(workspaceId: string): Promise<string | null> {
@@ -1673,24 +1650,6 @@ export async function runSpecCommand(
     customSpecRoot: options?.customSpecRoot ?? null,
     timeoutMs: options?.timeoutMs ?? null,
   });
-}
-
-export async function readAgentMd(workspaceId: string): Promise<AgentMdResponse> {
-  return fileRead("workspace", "agents", workspaceId);
-}
-
-export async function writeAgentMd(workspaceId: string, content: string): Promise<void> {
-  return fileWrite("workspace", "agents", content, workspaceId);
-}
-
-export type ClaudeMdResponse = TextFileResponse;
-
-export async function readClaudeMd(workspaceId: string): Promise<ClaudeMdResponse> {
-  return fileRead("workspace", "claude", workspaceId);
-}
-
-export async function writeClaudeMd(workspaceId: string, content: string): Promise<void> {
-  return fileWrite("workspace", "claude", content, workspaceId);
 }
 
 export async function listGitBranches(workspaceId: string): Promise<GitBranchListResponse> {
